requirements.txt
pandas
numpy
matplotlib
pyyaml
duka-dl
h2
